requirements.txt
# ==========================================
# CORE DEPENDENCIES
# ==========================================
Django>=5.1,<5.2
python-decouple>=3.8

# ==========================================
# FORMS & UI
# ==========================================
django-crispy-forms>=2.3
crispy-bootstrap5>=2024.2

# ==========================================
# API
# ==========================================
djangorestframework>=3.15.0
django-filter>=24.0
drf-spectacular>=0.27.0

# ==========================================
# DATA PROCESSING
# ==========================================
pandas>=2.2.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0

# ==========================================
# BIOMETRIC DEVICES
# ==========================================
pyzk>=0.9              # ZKTeco direct connection (TCP/UDP, port 4370)

# ==========================================
# PDF & IMAGES
# ==========================================
Pillow>=10.3.0          # ImageField, thumbnails
xhtml2pdf>=0.2.16       # Generación constancias/documentos PDF
pdfminer.six>=20221105  # Extracción texto PDF (bajas SUNAT, BLIQ liquidación S10)

# ==========================================
# DATABASE
# ==========================================
psycopg2-binary>=2.9.9
dj-database-url>=2.2.0

# ==========================================
# CACHE & PERFORMANCE
# ==========================================
django-redis>=5.4.0
redis>=5.0.0

# ==========================================
# ASYNC TASKS
# ==========================================
celery>=5.4.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.0

# ==========================================
# SECURITY
# ==========================================
django-cors-headers>=4.3.0

# ==========================================
# PRODUCTION SERVER
# ==========================================
gunicorn>=21.2.0
whitenoise>=6.6.0

# ==========================================
# MONITORING
# ==========================================
sentry-sdk>=2.0.0
python-json-logger>=3.0.0  # Para logging estructurado en JSON

# ==========================================
# DEVELOPMENT ONLY
# ==========================================
django-debug-toolbar>=4.3.0

# ==========================================
# TESTING
# ==========================================
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=24.0.0

# ==========================================
# CODE QUALITY
# ==========================================
ruff>=0.3.0
black>=24.0.0
isort>=5.13.0
